Add ultra scale images and other
my laptop camera is to small for face detection in the distance so the ultra scale x4, x8 can help, with the price of time in runtime.
my first page of code (where i try things so i can test them faster) will be not important. is more of a playground